requirements.txt
cloudpickle==1.3.0
Cython==0.29.23
glfw==2.1.0
gym==0.18.0
hydra-core==1.1.0.dev6
joblib==1.0.1
matplotlib==3.1.3
mlflow==1.17.0
mujoco-py==2.0.2.13
opencv-python-headless==4.5.2.54
optuna==2.7.0
stable-baselines3==1.0
torch==1.8.1
tqdm==4.60.0
comet_for_mlflow==0.1.2
